a8n: Split up a8n GraphQL resolver into multiple files
This does nothing more than splitting up the file. No identifiers
changed, no tests move, just more files.
